+++ mlir/unittests/Analysis/Presburger/SimplexTest.cpp
@@ -7,7 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "mlir/Analysis/Presburger/Simplex.h"
-#include "../utils.h"
+#include "../FACUtils.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
